+++ private/buf/cmd/buf/command/beta/registry/plugin/pluginpush/pluginpush.go
@@ -355,6 +355,7 @@ func createCuratedPluginRequest(
 		SpdxLicenseId:        pluginConfig.SPDXLicenseID,
 		LicenseUrl:           pluginConfig.LicenseURL,
 		Visibility:           visibility,
+		IntegrationGuideUrl:  pluginConfig.IntegrationGuideURL,
 	}, nil
 }